modularise styles
