+++ packages/core-ui/lib/buttons/ButtonHelp/index.scss
@@ -1,0 +1,26 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.btn-help {
+  @include styles.btn-layout;
+  @include styles.btn-spacing;
+  @include styles.btn-disabled;
+
+  background: var(--button-tertiary-background);
+  color: var(--text-color-secondary);
+  fill: var(--text-color-secondary);
+  border-radius: 50%;
+  padding: 0;
+  width: 1.35rem;
+  height: 1.35rem;
+
+  &.background-secondary {
+    background: var(--button-primary-background);
+  }
+
+  &:hover {
+    fill: var(--network-color-primary);
+  }
+}
