+++ packages/core-ui/lib/buttons/ButtonHelp/index.tsx
@@ -1,9 +1,10 @@
 /* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
 SPDX-License-Identifier: Apache-2.0 */
 
-import { InfoSVG } from "../svg/Info";
-import { ButtonCommonProps, ComponentBase } from "../types";
-import { valEmpty } from "../utils";
+import { InfoSVG } from "../../svg/Info";
+import { ButtonCommonProps, ComponentBase } from "../../types";
+import { valEmpty } from "../../utils";
+import "./index.scss";
 
 export type ButtonHelpProps = ComponentBase &
   ButtonCommonProps & {
