+++ packages/core-ui/lib/buttons/ButtonMono/index.scss
@@ -1,0 +1,29 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.btn-mono {
+  @include styles.btn-layout;
+  @include styles.btn-spacing;
+  @include styles.btn-icons;
+  @include styles.btn-disabled;
+
+  background: var(--text-color-primary);
+  border: 1px solid var(--text-color-primary);
+  border-radius: var(--button-border-radius-large);
+  color: var(--text-color-invert);
+
+  &.secondary-color {
+    border-color: var(--text-color-secondary);
+    background: var(--text-color-secondary);
+  }
+
+  &.sm {
+    @include styles.btn-small;
+  }
+
+  &.lg {
+    @include styles.btn-large;
+  }
+}
