+++ packages/core-ui/lib/buttons/ButtonOption/index.scss
@@ -1,0 +1,73 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.btn-option {
+  background: var(--button-primary-background);
+  transition: all var(--transition-duration);
+  padding: 1rem;
+  cursor: pointer;
+  margin-bottom: 1rem;
+  border-radius: 0.75rem;
+  display: flex;
+  align-items: center;
+  width: 100%;
+
+  &.content {
+    .details {
+      display: flex;
+
+      > h4 {
+        margin-left: 1rem;
+
+        span {
+          background: var(--background-primary);
+          border: 1px solid var(--border-primary-color);
+          margin-right: 0.5rem;
+          padding: 0.25rem 0.75rem;
+          border-radius: 1rem;
+        }
+      }
+    }
+  }
+
+  &:last-child {
+    margin-bottom: 0;
+  }
+
+  h3,
+  p {
+    text-align: left;
+    margin: 0;
+  }
+
+  h3 {
+    font-family: InterSemiBold, sans-serif;
+    margin-bottom: 0.5rem;
+  }
+
+  > div:last-child {
+    flex: 1;
+    display: flex;
+    justify-content: flex-end;
+  }
+
+  .icon {
+    margin-right: 0.5rem;
+  }
+
+  p {
+    color: var(--text-color-primary);
+    font-size: 1rem;
+  }
+
+  &:hover {
+    background: var(--button-hover-background);
+  }
+
+  &:disabled {
+    opacity: var(--opacity-disabled);
+    cursor: default;
+  }
+}
