+++ packages/core-ui/lib/buttons/ButtonOption/index.tsx
@@ -1,10 +1,11 @@
 /* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
 SPDX-License-Identifier: Apache-2.0 */
 
-import { ButtonCommonProps, ComponentBase } from "../types";
+import { ButtonCommonProps, ComponentBase } from "../../types";
 import { faChevronRight } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { valEmpty } from "../utils";
+import { valEmpty } from "../../utils";
+import "./index.scss";
 
 export type ButtonOptionProps = ComponentBase &
   ButtonCommonProps & {
