+++ packages/core-ui/lib/buttons/ButtonPrimaryInvert/index.scss
@@ -1,0 +1,28 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.btn-primary-invert {
+  @include styles.btn-layout;
+  @include styles.btn-spacing;
+  @include styles.btn-icons;
+  @include styles.btn-disabled;
+
+  border: 1px solid var(--network-color-stroke);
+  border-radius: var(--button-border-radius-large);
+  color: var(--network-color-stroke);
+
+  &.secondary-color {
+    border: 1px solid var(--network-color-secondary);
+    color: var(--network-color-secondary);
+  }
+
+  &.sm {
+    @include styles.btn-small;
+  }
+
+  &.lg {
+    @include styles.btn-large;
+  }
+}
