+++ packages/core-ui/lib/buttons/ButtonSubmit/index.tsx
@@ -1,48 +1,52 @@
 /* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
 SPDX-License-Identifier: Apache-2.0 */
 
-import { ButtonIconProps, ButtonCommonProps, ComponentBase } from "../types";
+import { ButtonIconProps, ButtonCommonProps, ComponentBase } from "../../types";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { valEmpty, valOr } from "../utils";
+import { valEmpty, valOr } from "../../utils";
 import { motion } from "framer-motion";
+import "./index.scss";
 
-export type ButtonMonoProps = ComponentBase &
+export type ButtonSubmitProps = ComponentBase &
   ButtonIconProps &
   ButtonCommonProps & {
-    // large button, small otherwise.
-    lg?: boolean;
+    // use secondary network color.
+    colorSecondary?: boolean;
     // button text.
     text: string;
+    // pulsing effect.
+    pulse?: boolean;
   };
 
 /**
- * @name ButtonMono
- * @description Monotone button style used within the main interface of dashboards.
+ * @name ButtonSubmit
+ * @description Submit button style used within modals to submit transactions.
  */
-export const ButtonMono = ({
+export const ButtonSubmit = ({
+  colorSecondary,
   disabled,
   grow,
   iconLeft,
   iconRight,
   iconTransform,
   onClick,
-  lg,
   marginLeft,
   marginRight,
   marginX,
   style,
   text,
-}: ButtonMonoProps) => (
+  pulse,
+}: ButtonSubmitProps) => (
   <motion.button
     whileHover={{ scale: !disabled ? 1.02 : 1 }}
     whileTap={{ scale: !disabled ? 0.98 : 1 }}
-    className={`btn-mono${valOr(lg, "lg", "sm")}${valEmpty(
-      grow,
-      "grow"
-    )}${valEmpty(marginRight, "m-right")}${valEmpty(
+    className={`btn-submit${valEmpty(
+      colorSecondary,
+      "secondary-color"
+    )}${valEmpty(grow, "grow")}${valEmpty(marginRight, "m-right")}${valEmpty(
       marginLeft,
       "m-left"
-    )}${valEmpty(marginX, "m-x")}`}
+    )}${valEmpty(marginX, "m-x")}${valEmpty(pulse, "pulse")}`}
     style={style}
     type="button"
     disabled={disabled}
