+++ packages/core-ui/lib/buttons/ButtonTab/index.scss
@@ -1,0 +1,47 @@
+/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
+SPDX-License-Identifier: Apache-2.0 */
+
+@use "../styles";
+
+.btn-tab {
+  @include styles.btn-layout;
+  @include styles.btn-disabled;
+
+  color: var(--text-color-primary);
+  transition:
+    opacity var(--transition-duration),
+    font-size var(--transition-duration);
+  font-size: 1.15rem;
+  padding: 0.65rem 1rem;
+  margin-bottom: 0.5rem;
+  margin-right: 0.75rem;
+  border-radius: 0.5rem;
+
+  &:last-child {
+    margin-right: 0;
+  }
+
+  &:hover {
+    opacity: 0.9;
+  }
+
+  &.active {
+    background: var(--button-tab-background);
+  }
+
+  > .badge {
+    @include styles.btn-small;
+
+    background: var(--network-color-secondary);
+    font-size: var(--button-font-size-small);
+    margin-left: var(--button-spacing-large);
+    border-radius: 0.4rem;
+    color: white;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    max-width: 3rem;
+    width: fit-content;
+    overflow: hidden;
+  }
+}
