+++ packages/core-ui/lib/core/styles.scss
@@ -1,9 +1,8 @@
 /* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
 SPDX-License-Identifier: Apache-2.0 */
 
-@use "variables" as v;
+@use "../styles/variables" as v;
 
-// padding of a page row component.
 @mixin page-padding {
   padding-left: 1.25rem;
   padding-right: 1.25rem;
