+++ packages/core-ui/lib/modal/styles.scss
@@ -1,7 +1,7 @@
 /* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
 SPDX-License-Identifier: Apache-2.0 */
 
-@use "variables" as v;
+@use "../styles/variables" as v;
 
 @mixin modal-position {
   position: fixed;
