+++ packages/core-ui/lib/styles/_variables.scss
