+++ packages/core-ui/lib/styles/index.scss
@@ -1,6 +1,12 @@
 /* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
 SPDX-License-Identifier: Apache-2.0 */
 
+/* Developer Notes:
+ * This file should be imported in the main library entry file. */
+
+/* UI theme variables. */
+@use "theme";
+
 /* Every element conforms to border-box by default. */
 * {
   box-sizing: border-box;
