--- packages/core-ui/styles/index.scss
@@ -1,24 +1,0 @@
-/* @license Copyright 2023 @paritytech/polkadot-dashboard-ui authors & contributors
-SPDX-License-Identifier: Apache-2.0 */
-
-/* Developer Notes:
- * This file is the entry point to which all of core-ui styles are built.
- * Additional files imported here will be a part of the final bundle. */
-
-/* Tag styling. */
-@use "tags";
-
-/* UI theme variables. */
-@use "theme";
-
-/* Core wrappers, tag styles and structural component styles. */
-@use "core";
-
-/* Button component styles. */
-@use "buttons";
-
-/* Modal component styles. */
-@use "modal";
-
-/* Hardware related component styles. */
-@use "hardware";
